+++ client/src/helper/urls.js
@@ -49,7 +49,7 @@ const urls = {
   mailboxMailto: (email) => `/?page=mailbox&mailto=${email}`,
   mailboxOldAttachment: (emailId, attachmentIndex) => `/?page=mailbox&a=dlattach&mid=${emailId}&i=${attachmentIndex}`,
   map: () => '/karte',
-  newsFromIT: () => '/content?sub=newsFromIT',
+  newsFromIT: () => 'https://foodsharing.freshdesk.com/support/solutions/folders/77000160479',
   vision: () => '/ueber-uns',
   partner: () => '/partner',
   passwordReset: () => '/?page=login&sub=passwordReset',